+++ src/components/LibraryEditForm.tsx
@@ -2,7 +2,13 @@ import * as React from "react";
 import EditableInput from "./EditableInput";
 import ProtocolFormField from "./ProtocolFormField";
 import { findDefault, clearForm } from "../utils/sharedFunctions";
-import { LibrariesData, LibraryData, LibrarySettingField, SettingData, AnnouncementData } from "../interfaces";
+import {
+  LibrariesData,
+  LibraryData,
+  LibrarySettingField,
+  SettingData,
+  AnnouncementData,
+} from "../interfaces";
 import { Panel, Form } from "library-simplified-reusable-components";
 import { FetchErrorData } from "opds-web-client/lib/interfaces";
 import PairedMenus from "./PairedMenus";
@@ -23,7 +29,10 @@ export interface LibraryEditFormProps {
 
 /** Form for editing a library's configuration, on the libraries tab of the
     system configuration page. */
-export default class LibraryEditForm extends React.Component<LibraryEditFormProps, {}> {
+export default class LibraryEditForm extends React.Component<
+  LibraryEditFormProps,
+  {}
+> {
   private nameRef = React.createRef<EditableInput>();
   private shortNameRef = React.createRef<EditableInput>();
   private settingRef = React.createRef<ProtocolFormField>();
@@ -33,13 +42,13 @@ export default class LibraryEditForm extends React.Component<LibraryEditFormProp
     this.submit = this.submit.bind(this);
   }
 
-  componentWillReceiveProps(nextProps) {
+  UNSAFE_componentWillReceiveProps(nextProps) {
     if (nextProps.responseBody && !nextProps.fetchError) {
       clearForm([
         this.nameRef.current,
         this.shortNameRef.current,
         this.settingRef.current,
-        this.announcementsRef.current
+        this.announcementsRef.current,
       ]);
     }
   }
@@ -48,7 +57,7 @@ export default class LibraryEditForm extends React.Component<LibraryEditFormProp
     // The server has assigned each setting a level from 1 to 3, indicating what level of permissions
     // the admin needs to have in order to be able to modify the item.  If the server has left the level blank,
     // the setting will be treated as if it is level 1 (editable by all admins).
-    return (setting.level && (setting.level > this.props.adminLevel));
+    return setting.level && setting.level > this.props.adminLevel;
   }
 
   render(): JSX.Element {
@@ -56,13 +65,18 @@ export default class LibraryEditForm extends React.Component<LibraryEditFormProp
     let otherFields = [];
 
     if (this.props.data && this.props.data.settings) {
-      basicInfo = this.props.data.settings.filter(setting => (setting.required || setting.category === "Basic Information"));
-      otherFields = this.props.data.settings.filter(setting => !(new Set(basicInfo).has(setting)));
+      basicInfo = this.props.data.settings.filter(
+        (setting) =>
+          setting.required || setting.category === "Basic Information"
+      );
+      otherFields = this.props.data.settings.filter(
+        (setting) => !new Set(basicInfo).has(setting)
+      );
     }
 
-    let categories = this.separateCategories(otherFields);
+    const categories = this.separateCategories(otherFields);
     const requiresSystemAdmin = 3;
-    let ref = (setting) => {
+    const ref = (setting) => {
       if (setting.key === "name") {
         return this.nameRef;
       } else if (setting.key === "short name") {
@@ -71,7 +85,7 @@ export default class LibraryEditForm extends React.Component<LibraryEditFormProp
         return this.settingRef;
       }
     };
-    let basicInfoPanel = (
+    const basicInfoPanel = (
       <Panel
         id="library-basic-info"
         headerText="Basic Information"
@@ -79,23 +93,27 @@ export default class LibraryEditForm extends React.Component<LibraryEditFormProp
         openByDefault={true}
         content={
           <fieldset>
-          <legend className="visuallyHidden">Basic Information</legend>
-          { basicInfo.map(setting =>
-            <ProtocolFormField
-              key={setting.key}
-              ref={ref(setting) as any}
-              setting={setting}
-              disabled={this.props.disabled}
-              value={this.props.item && this.props.item[setting.key] || this.props.item?.settings && this.props.item.settings[setting.key]}
-              default={findDefault(setting)}
-              error={this.props.error}
-              readOnly={this.adminNotAuthorized(setting)}
-            />
-          )
+            <legend className="visuallyHidden">Basic Information</legend>
+            {basicInfo.map((setting) => (
+              <ProtocolFormField
+                key={setting.key}
+                ref={ref(setting) as any}
+                setting={setting}
+                disabled={this.props.disabled}
+                value={
+                  (this.props.item && this.props.item[setting.key]) ||
+                  (this.props.item?.settings &&
+                    this.props.item.settings[setting.key])
+                }
+                default={findDefault(setting)}
+                error={this.props.error}
+                readOnly={this.adminNotAuthorized(setting)}
+              />
+            ))}
+          </fieldset>
         }
-        </fieldset>
-      }
-    />);
+      />
+    );
 
     return (
       <Form
@@ -105,29 +123,33 @@ export default class LibraryEditForm extends React.Component<LibraryEditFormProp
         disableButton={this.props.disabled}
         onSubmit={this.submit}
         content={[basicInfoPanel, this.renderForms(categories)]}
-      />);
+      />
+    );
   }
 
   separateCategories(nonRequiredFields: LibrarySettingField[]) {
-    let categories = {};
+    const categories = {};
     nonRequiredFields.forEach((setting) => {
-      categories[setting.category] = categories[setting.category] ? categories[setting.category].concat(setting) : [setting];
+      categories[setting.category] = categories[setting.category]
+        ? categories[setting.category].concat(setting)
+        : [setting];
     });
     return categories;
   }
 
-  renderForms(categories: {[key: string]: LibrarySettingField[]}) {
-    let forms = [];
-    let categoryNames = Object.keys(categories);
+  renderForms(categories: { [key: string]: LibrarySettingField[] }) {
+    const forms = [];
+    const categoryNames = Object.keys(categories);
     categoryNames.forEach((name) => {
-      let form =
+      const form = (
         <Panel
           id={`library-form-${name.replace(/\s/g, "-")}`}
           headerText={`${name} (Optional)`}
           onEnter={this.submit}
           key={name}
           content={this.renderFieldset(categories[name])}
-        />;
+        />
+      );
       forms.push(form);
     });
     return forms;
@@ -136,15 +158,15 @@ export default class LibraryEditForm extends React.Component<LibraryEditFormProp
   renderAnnouncements(setting: SettingData, value) {
     return (
       <AnnouncementsSection
-            setting={{...setting, ...{format: "date-range"}}}
-            value={value && JSON.parse(value)}
-            ref={this.announcementsRef}
+        setting={{ ...setting, ...{ format: "date-range" } }}
+        value={value && JSON.parse(value)}
+        ref={this.announcementsRef}
       />
     );
   }
 
   renderPairedMenus(setting: SettingData, fields: LibrarySettingField[]) {
-    let dropdownSetting = fields.find(x => x.key === setting.paired);
+    const dropdownSetting = fields.find((x) => x.key === setting.paired);
     // These dropdown settings have the :skip: attribute, so they'll get rendered only here;
     // when the iterator in renderFieldset gets to them, they'll get skipped over, not rendered a second time.
     return (
@@ -160,21 +182,21 @@ export default class LibraryEditForm extends React.Component<LibraryEditFormProp
   }
 
   renderFieldset(fields: SettingData[]) {
-    let formatSetting = (setting: SettingData) => {
+    const formatSetting = (setting: SettingData) => {
       if (setting.format === "narrow") {
         // We need to tell the ProtocolFormField to make an InputList with a menu that gets narrowed down.
-        return {...setting, ...{type: "menu"}};
+        return { ...setting, ...{ type: "menu" } };
       }
       return setting;
     };
-    let settingValue = (setting: SettingData) => {
+    const settingValue = (setting: SettingData) => {
       // Does the library already have a saved value for this setting?
       if (this.props.item?.settings) {
         return this.props.item.settings[setting.key];
       }
       return setting.default;
     };
-    let render = (setting) => {
+    const render = (setting) => {
       if (setting.paired) {
         return this.renderPairedMenus(setting, fields);
       } else if (setting.type === "announcements") {
@@ -190,7 +212,9 @@ export default class LibraryEditForm extends React.Component<LibraryEditFormProp
             value={settingValue(setting)}
             default={findDefault(setting)}
             error={this.props.error}
-            disableButton={this.adminNotAuthorized(setting) || this.props.disabled}
+            disableButton={
+              this.adminNotAuthorized(setting) || this.props.disabled
+            }
             readOnly={this.adminNotAuthorized(setting) || setting.readOnly}
           />
         );
@@ -199,17 +223,16 @@ export default class LibraryEditForm extends React.Component<LibraryEditFormProp
     return (
       <fieldset>
         <legend className="visuallyHidden">Additional Fields</legend>
-        { fields.map(setting => render(setting))}
+        {fields.map((setting) => render(setting))}
       </fieldset>
     );
   }
 
   submit(data: FormData) {
-    let announcements = this.announcementsRef.current.getValue();
-    let announcementList = [];
+    const announcements = this.announcementsRef.current.getValue();
+    const announcementList = [];
     announcements.forEach((a: AnnouncementData) => announcementList.push(a));
     data?.set("announcements", JSON.stringify(announcementList));
     this.props.save(data);
   }
-
 }